Fix [-Wsign-compare] warns in blocksdb code
These are the warning as produced by clang:

blockstorage/blockleveldb.cpp:121:35: warning: comparison of integers of different signs: 'int' and 'uint64_t' (aka 'unsigned long') [-Wsign-compare]
        if (pindexOldest->nHeight >= nLastBlockWeCanPrune)
            ~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~
blockstorage/blockstorage.cpp:404:34: warning: comparison of integers of different signs: 'uint64_t' (aka 'unsigned long') and 'int' [-Wsign-compare]
            if (lastFinishedFile <= loadedblockfile && index->nHeight > blockfiles[lastFinishedFile].nHeightLast)
                ~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~
blockstorage/blockstorage.cpp:404:71: warning: comparison of integers of different signs: 'int' and 'unsigned int' [-Wsign-compare]
            if (lastFinishedFile <= loadedblockfile && index->nHeight > blockfiles[lastFinishedFile].nHeightLast)
                                                       ~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
blockstorage/blockstorage.cpp:681:69: warning: comparison of integers of different signs: 'long' and 'size_t' (aka 'unsigned long') [-Wsign-compare]
            if (nCoinCacheMaxSize - nMaxCacheIncreaseSinceLastFlush > nTrimSize)

+++ src/blockstorage/blockleveldb.cpp
@@ -118,7 +118,7 @@ uint64_t FindFilesToPruneLevelDB(uint64_t nLastBlockWeCanPrune)
     CDBBatch undoBatch(*pblockundodb);
     while (nDBUsedSpace >= nPruneTarget && pindexOldest != nullptr)
     {
-        if (pindexOldest->nHeight >= nLastBlockWeCanPrune)
+        if (pindexOldest->nHeight >= (int)nLastBlockWeCanPrune)
         {
             break;
         }
